REF: Move stl related modules into single directory
