+++ statsmodels/tsa/seasonal.py
@@ -6,9 +6,9 @@ import pandas as pd
 from pandas.core.nanops import nanmean as pd_nanmean
 
 from statsmodels.tools.validation import PandasWrapper, array_like
-from statsmodels.tsa._stl import STL
+from statsmodels.tsa.stl._stl import STL
 from statsmodels.tsa.filters.filtertools import convolution_filter
-from statsmodels.tsa.mstl import MSTL
+from statsmodels.tsa.stl.mstl import MSTL
 from statsmodels.tsa.tsatools import freq_to_period
 
 __all__ = [
