+++ statsmodels/tsa/stl/__init__.py
@@ -1,0 +1,3 @@
+from statsmodels.tools._testing import PytestTester
+
+test = PytestTester()
