+++ statsmodels/tsa/stl/_stl.pyx
